Update web interface; finish digital control via MCP23008

+++ BoilerBotElectricalSystemMonitor/static/data/unpack.js
@@ -5,7 +5,7 @@ const MAPPING = {
     },
     '?': {
         size: 2,
-        converter: Uint8Array
+        converter: Uint16Array
     },
     'f': {
         size: 4,
